add page `options`, sort `requests` and `users` control

+++ app/Http/Controllers/AdminController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Admin;
 use App\Models\AdminQueue;
 use App\Models\Option;
 use App\Models\PcInfo;
@@ -26,20 +27,104 @@ class AdminController extends Controller
 			if (($request = AdminQueue::where('admin_id', Auth::user()->id)->first()) !== null) {
 				$distributed_request_id = $request->request_id;
 			}
-			$distributed_request = RequestInfo::where('id', $distributed_request_id)->first();
 			return view('admin.my', [
-				'my_requests' => RequestInfo::where('admin_id', Auth::user()->id)->where('status', 'В работе')->paginate(20),
-				'distributed_request' => $distributed_request
+				'my_requests' => RequestInfo::where('admin_id', Auth::user()->id)->orderBy('created_at', 'desc')->paginate(20),
+				'distributed_request' => RequestInfo::where('id', $distributed_request_id)->first(),
+				'type' => 'Все',
 			]);
 		}
 		abort(404);
 	}
 
+	public function my_completed()
+	{
+		if (Auth::user()->is_admin) {
+			$distributed_request_id = null;
+			if (($request = AdminQueue::where('admin_id', Auth::user()->id)->first()) !== null) {
+				$distributed_request_id = $request->request_id;
+			}
+			return view('admin.my', [
+				'my_requests' => RequestInfo::where('admin_id', Auth::user()->id)->where('status', 'Завершено')->orderBy('created_at', 'desc')->paginate(20),
+				'distributed_request' => RequestInfo::where('id', $distributed_request_id)->first(),
+				'type' => 'Завершённые',
+			]);
+		}
+		abort(404);
+	}
+
+	public function my_in_work()
+	{
+		if (Auth::user()->is_admin) {
+			$distributed_request_id = null;
+			if (($request = AdminQueue::where('admin_id', Auth::user()->id)->first()) !== null) {
+				$distributed_request_id = $request->request_id;
+			}
+			return view('admin.my', [
+				'my_requests' => RequestInfo::where('admin_id', Auth::user()->id)->where('status', 'В работе')->orderBy('created_at', 'desc')->paginate(20),
+				'distributed_request' => RequestInfo::where('id', $distributed_request_id)->first(),
+				'type' => 'В работе',
+			]);
+		}
+		abort(404);
+	}
+
+
 	public function requests()
 	{
 		if (Auth::user()->is_admin) {
 			$requests = RequestInfo::orderBy('created_at', 'desc')->paginate(20);
-			return view('admin.requests.index', compact('requests'));
+			return view('admin.requests.index', [
+				'requests' => $requests,
+				'type' => 'Все',
+			]);
+		}
+		abort(404);
+	}
+
+	public function requests_completed()
+	{
+		if (Auth::user()->is_admin) {
+			$requests = RequestInfo::where('status', 'Завершено')->orderBy('created_at', 'desc')->paginate(20);
+			return view('admin.requests.index', [
+				'requests' => $requests,
+				'type' => 'Завершённые',
+			]);
+		}
+		abort(404);
+	}
+
+	public function requests_in_work()
+	{
+		if (Auth::user()->is_admin) {
+			$requests = RequestInfo::where('status', 'В работе')->orderBy('created_at', 'desc')->paginate(20);
+			return view('admin.requests.index', [
+				'requests' => $requests,
+				'type' => 'В работе',
+			]);
+		}
+		abort(404);
+	}
+
+	public function requests_in_processing()
+	{
+		if (Auth::user()->is_admin) {
+			$requests = RequestInfo::where('status', 'В обработке')->orderBy('created_at', 'desc')->paginate(20);
+			return view('admin.requests.index', [
+				'requests' => $requests,
+				'type' => 'В обработке',
+			]);
+		}
+		abort(404);
+	}
+
+	public function requests_canceled()
+	{
+		if (Auth::user()->is_admin) {
+			$requests = RequestInfo::where('status', 'Отменено')->orderBy('created_at', 'desc')->paginate(20);
+			return view('admin.requests.index', [
+				'requests' => $requests,
+				'type' => 'Отменённые',
+			]);
 		}
 		abort(404);
 	}
@@ -71,17 +156,75 @@ class AdminController extends Controller
 	public function user_show($user_id)
 	{
 		if (Auth::user()->is_admin) {
-			if (User::findOrFail($user_id)->is_admin) {
+			$user = User::findOrFail($user_id);
+			if ($user->is_admin) {
 				return view('admin.users.show', [
-					'user' => User::findOrFail($user_id),
+					'user' => $user,
 					'created' => count(RequestInfo::where('user_id', $user_id)->get()),
 					'done' => count(RequestInfo::where('admin_id', $user_id)->where('status', 'Завершено')->get()),
 				]);
 			}
 			return view('admin.users.show', [
-				'user' => User::findOrFail($user_id),
+				'user' => $user,
 				'created' => count(RequestInfo::where('user_id', $user_id)->get()),
 			]);
+		}
+		abort(404);
+	}
+
+	public function change_status($user_id, $status)
+	{
+		if (Auth::user()->admin->is_master) {
+			$user = User::findOrFail($user_id);
+			switch ($status) {
+				case 'admin':
+					if (!$user->is_admin) {
+						$user->is_admin = true;
+						$user->save();
+						Admin::create([
+							'user_id' => $user_id,
+						]);
+					}
+					break;
+				case 'master':
+					if ($user->is_admin && !$user->admin->is_master) {
+						$admin = $user->admin;
+						$admin->is_master = true;
+						$admin->save();
+					}
+					break;
+				case 'downgrade':
+					if ($user->admin->is_master) {
+						$admin = $user->admin;
+						$admin->is_master = false;
+						$admin->save();
+					} else {
+						if ($user->is_admin) {
+							$user->is_admin = false;
+							$user->save();
+							$admin = $user->admin;
+							$admin->delete();
+						}
+					}
+					break;
+				default:
+					abort(404);
+					break;
+			}
+			return redirect('/admin/users/' . $user_id);
+		}
+		abort(404);
+	}
+
+	public function user_destroy(User $user)
+	{
+		if (Auth::user()->is_admin && Auth::user()->admin->is_master) {
+			if ($user->is_admin) {
+				$admin = $user->admin;
+				$admin->delete();
+			}
+			$user->delete();
+			return redirect('/admin/users');
 		}
 		abort(404);
 	}
@@ -109,7 +252,7 @@ class AdminController extends Controller
 	public function cancel(RequestInfo $request)
 	{
 		if (Auth::user()->is_admin) {
-			if ($request->status == 'В обработке' || ($request->status == 'В работе' &&  $request->admin_id == Auth::user()->id)) {
+			if ($request->status == 'В обработке' || ($request->status == 'В работе' &&  ($request->admin_id == Auth::user()->id || Auth::user()->admin->is_master))) {
 				$request->status = 'Отменено';
 				$request->time_remaining = NULL;
 				$request->closed_at = Carbon::now();
@@ -231,16 +374,40 @@ class AdminController extends Controller
 		abort(404);
 	}
 
-	public function destroy(RequestInfo $request)
+	public function request_destroy(RequestInfo $request)
 	{
-		if (Auth::user()->is_admin) {
-			if ($request->status != 'В работе' || $request->admin_id == Auth::user()->id) {
+		if (Auth::user()->is_admin && Auth::user()->admin->is_master) {
+			if ($request->status != 'В работе' || $request->admin_id == Auth::user()->id || Auth::user()->admin->is_master) {
 				//ДОБАВИТЬ
 				//Если есть зависимость (подробная информация) -> удаляем зависимую запись
 				$request->delete();
 				return redirect('/admin/requests');
 			}
 			return redirect('/admin/requests/' . $request->id);
+		}
+		abort(404);
+	}
+
+	public function options()
+	{
+		if (Auth::user()->is_admin) {
+			$admin = Auth::user()->admin;
+			$options = Option::first();
+			return view('admin.options', [
+				'admin' => $admin,
+				'options' => $options,
+			]);
+		}
+		abort(404);
+	}
+
+	public function recommendation(Request $request)
+	{
+		if (Auth::user()->is_admin) {
+			$admin = Auth::user()->admin;
+			$admin->get_recommendation = $request->status;
+			$admin->save();
+			return response()->json(['success' => 'Recommendation changed successfully.']);
 		}
 		abort(404);
 	}
